Moved Utils
