+++ src/Agent/NewRelic/Agent/Core/BrowserMonitoring/BrowserMonitoringStreamInjector.cs
@@ -1,7 +1,7 @@
 // Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-using NewRelic.Agent.Core.Utils;
+using NewRelic.Agent.Core.Utilities;
 using NewRelic.Core;
 using NewRelic.Core.Logging;
 using System;
